fix(ui): derive the email-status query instead of writing it out
A hand-written subscription body that names an attribute slightly
differently matches nothing, and a subscription matching nothing is
indistinguishable from a service that never answered — which is what the
entry screen looked like: the button never relabelled and the lookup
appeared to hang.

The query builder moves out of the test-gated helpers into
`tonk_worker::query`, so the account panel, the share dialog, and the
tests that ask what an address answered all derive it from the concept.
